sql: split parser.Datum into Type and Datum
Currently we use Datum to represent both types and values, with naming
conventions that indicate the intended use. Having separate types is
clearer and will aid future changes to the type system.

Remove DPlaceholder -- placeholders are variables, not values. To
compensate, Placeholder becomes a TypedExpr with type TPlaceholder.

Since we're touching adjacent code anyway, fix the Size computation for
DDecimal.