sql/parser: fix serialization for DInterval
An interval is represented in SQL with syntax like `INTERVAL '5m'`. But when we
serialize `DInterval`s, they print out as just `5m`, which results in
invalid expressions.

This change fixes the serialization and provides a new function `ValueAsString`
which is used when casting to string (in that case we do want just the value as
a string).