sql: plumb contexts high and wide around sql
The objective is to be more explicit about context passing: don't store
the context in some planNodes and get rid of planner.Ctx().

- a ctx is added to important planNode methods
- don't bind a Context in BoundAccount*

references #7992

*a BoundAccount is a MemoryAccount bound to a MemoryMonitor. When
creating a BoundAcc, we used to bind a context to be used by the Acc
throughout its lifetime. This means that the context had to stay "valid"
throughout (e.g. the span in the ctx, if any, had to stay open). This is
conceptually a problem because it means you can't open an acc in one
place and close it in a completely different place (and also diminishes
the utility of contexts, as Account operations were not logged in a
context corresponding to the function actually performing them). In
particular, this was a problem for accounts created for RowContainers
and the way how EXPLAIN(TRACE) temporarily hijacks the session's
context: a container's account could easily be bound to one of these
temporarily hijacked contexts, and closed after the corresponding span
had already been Finish()ed.
This patch plumbs through a Context to the accounts, through various
cascading layers.