sql: improve OID type representation
Previously, the Postgres OID type was represented as a DOidTypeWrapper
around a DInt. This was a shortcut implementation that satisfied the
basic requirements of OIDs and their cast behaviors, but was missing a
critical piece: representation as either an integer, the underlying OID,
or a string, the name of the entity that the OID represents.

The OIDTypeWrapper implementation is now removed. It's replaced by a
fully-fledged OID datum, DOid, which carries the underlying OID integer,
the type of the OID (regproc, regclass, regtype, etc), and the optional
entity name that it is resolved to.

This opens the door to having OID columns in pg_catalog tables appear as
strings when selected but still properly join against other OID columns.