sql/parser: Add EvalContext as parameter to Datum.Compare
This requires a significant propagation of `*parser.EvalContext`s,
because this is now a required parameter of `Datum.Compare`. In most
cases, this is actually a good thing, because it gives us more flexibility
in using caches and session variables in these comparison methods.

This is necessary for a later commit which merges `Datum.Compare` and
`CmpOps` functions.