sql/parser: Permit placeholders for OID values
Part of #14245.

Previously the `tOID` type claimed that it was ambiguous through its
`IsAmbiguous` method. This meant that a `placeholderTypeAmbiguityError`
would be thrown when issuing queries like `SELECT 1::oid = $1`. This was
incorrect, as an OID type is fully defined.

This change fixes this, and also improves the `tOid` type slightly. It
gets rid of its redundant `typeName` field. Additionally, it improves
the type's `Equivalent` and `FamilyEqual` methods, which would have
given us issues if one of the "reg oid" variants (`typeRegClass`,
`typeRegProc`, etc.) ever entered our type system.