sql, encoding: update apd, with NaN, Inf support
Sync to most recent apd branch that has NaN, Infinity, and negative
zero support. The Traps have been left the same so that errors are,
in general, produced just as they were before.

The notable difference is `power(0, -1)`, which produces Infinity
(and is perhaps the only known way to produce Infinity using
arithmetic). This is because the GDA spec tests consider this a valid
operation. It sadly differs from postgres. However, these differences
from postgres are already present and inconsistent, so it's no worse
than we already were, and will be fixed sometime soon.

This change also consolidates decimal string parsing into a common
method in the parser package which takes care of converting the 4
various forms of NaN into a canonical form (just NaN).