sql/parser: Treat StrVal and NumVal the same during overload resolution
Fixes #14527.

Previously, our untyped literal handling was inconsistent for string
literals. We made the decision a while ago to treat string literals as
resolved string types during overload resolution. This decision was made
because it simplified certain logic, avoided ambiguities in a few cases,
and allowed a certain class of errors to be more specific. However, this
approach led to inconsistencies in untyped string constant resolution (see
the linked issue).

This change changes overload resolution and homogeneous type checking to
give `StrVal` constants the same treatment as `NumVal` constants. The
change addresses inconsistencies like the related issue. The cost of
this is that it introduces a limitation that we already had for numeric
constants, which materializes in ambiguities in constant coercion between
parent expressions (see change to `pgwire_test.go`).

This change makes a few errors worse, like the error reported for `'a' + 0`.
Previously we would return and `unsupported binary operator` error. Instead, we
will now return a `could not parse 'a' as type date` error, because type
checking will try to use our `<date> + <int>` overload. An unexpected
effect is that it also improves some errors, like the one for
`COALESCE(1, 2, 3, 4, '5')`. Now, instead of throwing an error that `1
is not a string`, we'll read left to right and throw an error that `'5'
is not an int`.