sql: track memory used by builtins
This commit introduces memory tracking for various builtin functions.
The functions which allocate memory and thus currently track it are
LOWER, UPPER, CONCAT, CONCAT_WS, REPEAT, REVERSE, REPLACE, TRANSLATE,
REGEXP_REPLACE, and INITCAP.

It's worth noting that many functions, such as SPLIT_PART, that return
strings return a slice of the input, and so do not need to track that
memory.

This is related to #15295, but is a simplified version of a more
complete solution.

A drawback of the current use of ActiveMemAcc is that it's only set in
the executor and tracks allocations on a per-output-row basis. This
means that nodes that take in multiple input rows per output row can
overcount their memory. However, with the current setup the only way to
address this would be to swap out the ActiveMemAcc on the nodes which do
that prior to them grabbing their next row, and then replacing the old
one after. A more correct solution might be to create a new function
which pushes a fresh memory account onto a stack, evaluates the next
row, and then closes that account, but I've gone with this for now since
it's simpler and errs on the safe side of memory accounting. If we move
to this more sophisticated approach I also think a Grow-like method on
EvalContext might be more appropriate.

In addition, the current process of constant folding discards any errors
observed while evaluating expressions, including memory errors. This
is mostly inconsequential