sql: rollback disallowing statements following schema change
rolls back #15339 #14368 for the most part and #14619 partially.

We used a very big hammer to prevent some statements from
following a schema change. This was done to reduce surprises
but it ended up introducing more surprises. ORMs and schema
migration tools depend on multiple schema changes being
staged in the same transaction.

fixes #15269