sql: improve string formatting for postgres-compat
This change introduces a new method of formatting strings within arrays
which is more compatible with how Postgres formats them. It's also a
slightly simplified version of what Postgres does. The primary
difference is that we quote every string, and we use double quotes
instead of single quotes.

These changes are required for the node.js postgres driver.

I added tests to the python and node acceptance tests.
The ruby pg gem doesn't seem to support arrays at all, and adding a java
one is blocked on #17140.

I ran a quick benchmark comparing the two functions:

```
func BenchmarkEncodeSQLString(b *testing.B) {
	str := strings.Repeat("foo", 10000)
	b.Run("old version", func(b *testing.B) {
		for i := 0; i < b.N; i++ {
			encodeSQLStringWithFlags(bytes.NewBuffer(nil), str, FmtBareStrings)
		}
	})
	b.Run("new version", func(b *testing.B) {
		for i := 0; i < b.N; i++ {
			encodeSQLStringInsideArray(bytes.NewBuffer(nil), str)
		}
	})
}
```

and the results are that the new version is substantially faster, though
this isn't entirely fair since there are no escaped characters in the
benchmark, so the old version does a lot of unnecessary work. It is
reassuring that there isn't much, if anything, lost due to inserting
into the buffer char-by-char using WriteByte instead of using
WriteString.:

BenchmarkEncodeSQLString/old_version-8  2000 1035068 ns/op
BenchmarkEncodeSQLString/new_version-8  5000 304645 ns/op