sql: make SET and SET CLUSTER SETTING support the prepare protocol
This patch ensures that a client can use placeholders in the value
field of SET and SET CLUSTER SETTING statements.

For example:

```sql
PREPARE lateset(string) AS SET application_name = $1
EXECUTE lateset('myapp')
```

And there was much rejoicing!

As a side effect of supporting this, these two statements now also
support values defined by subqueries. For example:

```sql
SET application_name = (SELECT 'my' || 'app')
```

Note: this patch does *not* support placeholders on the LHS of the
assignment. This is because the type required for the RHS depends on
the value of the LHS (which setting is being modified). This is the
same reason why we do not currently support placeholders for table
names in INSERT, etc.